submit abc100_c

+++ abc/100/c.cpp
@@ -9,7 +9,7 @@ int main() {
   int n;
   cin >> n;
 
-  vector<int64_t> a(n);
+  vector<int> a(n);
   for (int i = 0; i < n; ++i) {
     cin >> a[i];
   }
